pyproject.toml
[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = "An image classification model pipeline, ultimately to classify dog locations within a room."
authors = ["pitta-bread <peterbeardsmore@hotmail.co.uk>"]
readme = "README.md"
packages = [{include = "ml_pipeline"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
tensorflow = "2.10.1"
keras-tuner = "^1.3.3"
graphviz = "^0.20.1"
tensorflow-io-gcs-filesystem = "0.31.0"
pydot = "^1.4.2"
pydotplus = "^2.0.2"
scikit-learn = "^1.2.2"
google-cloud-storage = "^2.8.0"
matplotlib = "^3.7.1"
flask = "^2.2.3"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
